+++ service/storage/impl/src/main/java/com/emc/pravega/service/storage/impl/distributedlog/LogClient.java
@@ -81,7 +81,7 @@ class LogClient implements AutoCloseable {
 
     @Override
     public void close() {
-        int traceId = LoggerHelpers.traceEnter(log, this.traceObjectId, "close", this.closed);
+        long traceId = LoggerHelpers.traceEnter(log, this.traceObjectId, "close", this.closed);
         if (!this.closed) {
             ArrayList<LogHandle> handlesToClose;
             synchronized (this.handles) {
@@ -121,7 +121,7 @@ class LogClient implements AutoCloseable {
      *                                 the failure reason.
      */
     public void initialize() throws DurableDataLogException {
-        int traceId = LoggerHelpers.traceEnter(log, this.traceObjectId, "initialize");
+        long traceId = LoggerHelpers.traceEnter(log, this.traceObjectId, "initialize");
         Exceptions.checkNotClosed(this.closed, this);
         Preconditions.checkState(this.namespace == null, "LogClient is already initialized.");
